darwin: properly handle 16-bit UUIDs for service and characteristics in the unique format used by macOS
